initial filtering working

+++ src/pyfilter/models/_discretization.py
@@ -11,5 +11,5 @@ def expm_discretizer(A: FloatArray, dt: FloatArray) -> FloatArray:
 
 def euler_discretizer(A: FloatArray, dt: FloatArray) -> FloatArray:
     """First-order Euler: Phi = I + A * dt."""
-    eye = np.eye(A.shape[-1])
+    eye = np.eye(A.shape[-1], dtype=A.dtype)
     return eye + A * dt
